requirements.txt
fastapi
uvicorn
SQLAlchemy
psycopg2-binary
python-dotenv
loguru
pydantic_settings